fix(repo): fix commitlint header-max-length
This rule was originally set as warning, but editor (namely, Emacs)
ignores warning and just make a commit, which is not preferable
behaviour. Therefore I change this rule to emit error.

Originally maximum length of header was 50, because I thought 50/72
rule is nice one. However, it turns out that this this rule is too
strict to apply.
I update the length to 72 because GitHub accepts 72 characters long
commit header, in the other words, GitHub truncates commit header
including more than 72 characters.